backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
requests
python-dotenv
openai
supabase
pgvector
